Unify the PostSchedule component between site and post editors
