+++ packages/e2e-tests/specs/editor/various/datepicker.test.js
@@ -56,7 +56,7 @@ function formatDatePickerValues(
 
 async function getPublishingDate() {
 	return page.$eval(
-		'.edit-post-post-schedule__toggle',
+		'.editor-post-schedule__dialog-toggle',
 		( dateLabel ) => dateLabel.textContent
 	);
 }
@@ -83,7 +83,7 @@ describe.each( [ [ 'UTC-10' ], [ 'UTC' ], [ 'UTC+10' ] ] )(
 
 		it( 'should show the publishing date if the date is in the past', async () => {
 			// Open the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			// Change the publishing date to a year in the past.
 			await page.click( '.components-datetime__time-field-year' );
@@ -91,7 +91,7 @@ describe.each( [ [ 'UTC-10' ], [ 'UTC' ], [ 'UTC+10' ] ] )(
 			const datePickerValues = await getDatePickerValues();
 
 			// Close the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			const publishingDate = await getPublishingDate();
 
@@ -102,7 +102,7 @@ describe.each( [ [ 'UTC-10' ], [ 'UTC' ], [ 'UTC+10' ] ] )(
 
 		it( 'should show the publishing date if the date is in the future', async () => {
 			// Open the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			// Change the publishing date to a year in the future.
 			await page.click( '.components-datetime__time-field-year' );
@@ -110,7 +110,7 @@ describe.each( [ [ 'UTC-10' ], [ 'UTC' ], [ 'UTC+10' ] ] )(
 			const datePickerValues = await getDatePickerValues();
 
 			// Close the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			const publishingDate = await getPublishingDate();
 
@@ -123,17 +123,17 @@ describe.each( [ [ 'UTC-10' ], [ 'UTC' ], [ 'UTC+10' ] ] )(
 
 		it( `should show the publishing date as "Immediately" if the date is cleared`, async () => {
 			// Open the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			// Change the publishing date to a year in the future.
 			await page.click( '.components-datetime__time-field-year' );
 			await page.keyboard.press( 'ArrowUp' );
 
 			// Close the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			// Open the datepicker.
-			await page.click( '.edit-post-post-schedule__toggle' );
+			await page.click( '.editor-post-schedule__dialog-toggle' );
 
 			// Clear the date.
 			await page.click(
