--- packages/edit-post/src/components/sidebar/post-schedule/style.scss
@@ -1,36 +1,0 @@
-.edit-post-post-schedule {
-	width: 100%;
-	position: relative;
-	justify-content: flex-start;
-	align-items: flex-start;
-
-	span {
-		display: block;
-		width: 30%;
-		margin-right: 8px;
-	}
-}
-
-.edit-post-post-schedule__dropdown {
-	width: 70%;
-}
-
-.components-button.edit-post-post-schedule__toggle {
-	text-align: left;
-	white-space: normal;
-	height: auto;
-
-	// This span is added by the Popover in Tooltip when no anchor is
-	// provided. We set its width to 0 so that it does not cause the button text
-	// to wrap to a new line when displaying the tooltip. A better fix would be
-	// to pass anchorRef and avoid the need for a span alltogether, which is
-	// what this PR allows us to do:
-	// https://github.com/WordPress/gutenberg/pull/41268.
-	span {
-		width: 0;
-	}
-}
-
-.edit-post-post-schedule__dialog .block-editor-publish-date-time-picker {
-	margin: $grid-unit-10;
-}
