+++ packages/edit-post/src/components/sidebar/post-status/index.js
@@ -8,6 +8,7 @@ import {
 } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
 import {
+	PostSchedulePanel,
 	PostSwitchToDraftButton,
 	PostSyncStatus,
 	PostURLPanel,
@@ -18,7 +19,6 @@ import {
  */
 import PostVisibility from '../post-visibility';
 import PostTrash from '../post-trash';
-import PostSchedule from '../post-schedule';
 import PostSticky from '../post-sticky';
 import PostAuthor from '../post-author';
 import PostSlug from '../post-slug';
@@ -61,7 +61,7 @@ export default function PostStatus() {
 				{ ( fills ) => (
 					<>
 						<PostVisibility />
-						<PostSchedule />
+						<PostSchedulePanel />
 						<PostTemplate />
 						<PostURLPanel />
 						<PostSticky />
