+++ packages/editor/src/components/index.js
@@ -50,6 +50,7 @@ export {
 	default as PostScheduleLabel,
 	usePostScheduleLabel,
 } from './post-schedule/label';
+export { default as PostSchedulePanel } from './post-schedule/panel';
 export { default as PostSlug } from './post-slug';
 export { default as PostSlugCheck } from './post-slug/check';
 export { default as PostSticky } from './post-sticky';
