+++ packages/editor/src/components/post-schedule/style.scss
@@ -1,0 +1,21 @@
+.editor-post-schedule__panel {
+	width: 100%;
+	justify-content: flex-start;
+	align-items: flex-start;
+
+	span {
+		display: block;
+		width: 30%;
+	}
+}
+
+.editor-post-schedule__panel-dropdown {
+	width: 70%;
+}
+
+.editor-post-schedule__dialog {
+	.components-popover__content {
+		min-width: 320px;
+		padding: $grid-unit-20;
+	}
+}
